[Misc] Migrate the LinkChecker functional test to the JUnit5 Docker test framework
Convert LinkCheckerTest (JUnit4, xwiki-platform-test-ui, packager plugin) to the JUnit5 Docker
framework (@UITest) in a new xwiki-platform-linkchecker-test-docker module, and remove the old
packager-based xwiki-platform-linkchecker-test-tests module. The single test (linkChecker) keeps its
original assertion on the "External Links" livetable and gains one on the reported link state (404).
That new assertion fails if StubHTTPChecker ever stops being the registered HTTPChecker: the real
implementation answers 0 for an unresolvable host and the livetable row is displayed whatever the
response code, so a bypassed stub would otherwise go unnoticed.

The StubHTTPChecker component (which keeps the test off the network) moves into the new module's
src/main. Since it ships in WEB-INF/classes and is parsed by the component annotation loader at WAR
startup, the HTTPChecker role it implements must be in WEB-INF/lib at that point; that role lives in
the linkchecker rendering transformation, which is not part of the minimal test WAR, so it is added
via @UITest(extraJARs=...). The linkchecker transformation is activated by appending it to
rendering.transformations in xwiki.properties.
