+++ xwiki-platform-core/xwiki-platform-linkchecker/xwiki-platform-linkchecker-test/pom.xml
@@ -41,6 +41,13 @@
   </properties>
   <modules>
     <module>xwiki-platform-linkchecker-test-pageobjects</module>
-    <module>xwiki-platform-linkchecker-test-tests</module>
   </modules>
+  <profiles>
+    <profile>
+      <id>docker</id>
+      <modules>
+        <module>xwiki-platform-linkchecker-test-docker</module>
+      </modules>
+    </profile>
+  </profiles>
 </project>
