+++ xwiki-platform-core/xwiki-platform-linkchecker/xwiki-platform-linkchecker-test/xwiki-platform-linkchecker-test-docker/src/main/java/org/xwiki/test/linkchecker/internal/StubHTTPChecker.java
